refactor: fetch data directly from db
